+++ ricaun.Nuke/Components/IGitRelease.cs
@@ -1,8 +1,9 @@
 ﻿using Nuke.Common;
 using Nuke.Common.Git;
 using Nuke.Common.IO;
+using Nuke.Common.ProjectModel;
 using Nuke.Common.Tools.GitHub;
-using Octokit;
+using Nuke.Common.Utilities.Collections;
 using ricaun.Nuke.Extensions;
 using System;
 using System.IO;
@@ -28,44 +29,62 @@ namespace ricaun.Nuke.Components
             {
                 var project = MainProject;
 
-                if (Directory.Exists(ReleaseDirectory) == false)
+                if (project.IsVersionPreRelease())
                 {
-                    Serilog.Log.Warning($"Release Directory not Found: {ReleaseDirectory}");
-                    return;
+                    ReportSummary(_ => _.AddPair("Prerelease", project.GetInformationalVersion()));
+                    var errorMessage = $"The project {project.Name} is a pre-release";
+                    Serilog.Log.Error(errorMessage);
+                    throw new Exception(errorMessage);
                 }
 
-                Serilog.Log.Information($"GitHubTasks.GitHubClient: {Solution.Name}");
-
-                GitHubTasks.GitHubClient = new GitHubClient(new ProductHeaderValue(Solution.Name))
-                {
-                    Credentials = new Credentials(GitHubToken)
-                };
-
-                var gitHubName = GitRepository.GetGitHubName();
-                var gitHubOwner = GitRepository.GetGitHubOwner();
-
-                var releaseFiles = Globbing.GlobFiles(ReleaseDirectory, "*.zip");
-                var version = project.GetInformationalVersion();
-
-                Serilog.Log.Information($"GitHubTasks.CheckTags: {gitHubOwner} {gitHubName} {version}");
-
-                if (GitHubExtension.CheckTags(gitHubOwner, gitHubName, version))
-                {
-                    Serilog.Log.Warning($"The repository already contains a Release with the tag: {version}");
-                    return;
-                }
-
-                var newRelease = new NewRelease(version)
-                {
-                    Name = version,
-                    Body = GetReleaseNotes(),
-                    Draft = true,
-                    TargetCommitish = GitVersion.Sha
-                };
-
-                var draft = GitHubExtension.CreatedDraft(gitHubOwner, gitHubName, newRelease);
-                GitHubExtension.UploadReleaseFiles(draft, releaseFiles);
-                GitHubExtension.ReleaseDraft(gitHubOwner, gitHubName, draft);
+                ReleaseGithubProject(project);
             });
+
+        /// <summary>
+        /// Release Github project with release notes
+        /// </summary>
+        /// <param name="project"></param>
+        /// <param name="releaseAsPrerelease"></param>
+        void ReleaseGithubProject(Project project, bool releaseAsPrerelease = false)
+        {
+            if (Directory.Exists(ReleaseDirectory) == false)
+            {
+                Serilog.Log.Warning($"Release Directory not Found: {ReleaseDirectory}");
+                return;
+            }
+
+            Serilog.Log.Information($"GitHubTasks.GitHubClient: {Solution.Name}");
+
+            GitHubTasks.GitHubClient = new Octokit.GitHubClient(new Octokit.ProductHeaderValue(Solution.Name))
+            {
+                Credentials = new Octokit.Credentials(GitHubToken)
+            };
+
+            var gitHubName = GitRepository.GetGitHubName();
+            var gitHubOwner = GitRepository.GetGitHubOwner();
+
+            var releaseFiles = Globbing.GlobFiles(ReleaseDirectory, "*.zip");
+            var version = project.GetInformationalVersion();
+
+            Serilog.Log.Information($"GitHubTasks.CheckTags: {gitHubOwner} {gitHubName} {version}");
+
+            if (GitHubExtension.CheckTags(gitHubOwner, gitHubName, version))
+            {
+                Serilog.Log.Warning($"The repository already contains a Release with the tag: {version}");
+                return;
+            }
+
+            var newRelease = new Octokit.NewRelease(version)
+            {
+                Name = version,
+                Body = GetReleaseNotes(),
+                Draft = true,
+                TargetCommitish = GitVersion.Sha
+            };
+
+            var draft = GitHubExtension.CreatedDraft(gitHubOwner, gitHubName, newRelease);
+            GitHubExtension.UploadReleaseFiles(draft, releaseFiles);
+            GitHubExtension.ReleaseDraft(gitHubOwner, gitHubName, draft, releaseAsPrerelease);
+        }
     }
 }