+++ CHANGELOG.md
@@ -4,6 +4,21 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [1.7.0] / 2023-10-04
+### Features
+- Prerelease feature for Github and Nuget.
+### Updated
+- Update `AssemblyExtension` to fix `GetVersion` prerelease key work.
+- Update `ITest` to runs before `Release`
+- Update `IGitRelease` with custom method to create release and pre-release.
+- Update `GitHubExtension` to release or pre-release.
+- Update order of `Targets`
+- Update `IGitRelease` to fail if MainProject is prerelease.
+- Update `IHazPack` with base package methods.
+### Added
+- Add `IGitPreRelease` with custom method to create pre-release when version is pre-release.
+- Add `IPrePack` with custom method to create pre-release when version is pre-release.
+
 ## [1.6.1] / 2023-10-02
 ### Fixed
 - Fix test file with multi target framework (Fix: #60)
@@ -294,6 +309,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - First Release
 
 [vNext]: ../../compare/1.0.0...HEAD
+[1.7.0]: ../../compare/1.6.1...1.7.0
 [1.6.1]: ../../compare/1.6.0...1.6.1
 [1.6.0]: ../../compare/1.5.2...1.6.0
 [1.5.2]: ../../compare/1.5.1...1.5.2